Ensure providers and aliases get added automatically
Moved the configure function call of the providerManager outside of the ConfigurablePackageInterface check so packages don’t have to implement this interface for their provider and alias to get added to the `config/app.php` file automatically. If a package doesn’t specifically sets the provider or alias it won’t get added anyway.